[caliper] remove unnecessary reference to the hierachical format
